+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * refactor: remove `Countable` from minimal classes
 * refactor: rename `RestrictedCountStrategy` to `DisabledCountStrategy`
 * feat: `$offset` parameter type widening for `ArrayAccess` methods
+* fix(`DefaultKeyTransformer`): only convert `Stringable` to string for now.
 
 ## 0.7.0